requirements.txt
httplib2==0.20.4
matplotlib==3.6.0
seaborn==0.12.0
serial==0.0.97
sklearn-model==0.0.6
pandas==1.5.0
python-dotenv==0.21.0